refactor(yki): refactor folders and naming
 - no more responses nor entities -folders (since there would be only one file per folder)
 - rename YkiSuotritusReponse to SolkiSuoritusResponse
	this is because the external system is solki, and not really related to yki.
	It (solki) could be event it's own domain.
