+++ server/src/main/kotlin/fi/oph/kitu/yki/YkiRepository.kt
@@ -1,6 +1,5 @@
 package fi.oph.kitu.yki
 
-import fi.oph.kitu.yki.entities.YkiSuoritusEntity
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
